requirements.txt
adjustText==0.7.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
cfgv==3.3.1
cycler==0.11.0
decorator==5.1.0
default==0.1.2
distlib==0.3.4
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.3
environ==1.0
filelock==3.4.0
fonttools==4.28.3
identify==2.4.0
joblib==1.1.0
kiwisolver==1.3.2
many==0.6.8
matplotlib==3.5.1
nodeenv==1.6.0
numpy==1.21.4
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.16.0
pre-commit-hooks==4.0.1
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
self==2020.12.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
statsmodels==0.12.2
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
virtualenv==20.10.0
